adding source and target link removal
when you hover over a sheet link (target or source) an 'unlink' icon
appears. Clicking on this removes the link (note this happens on bote the source
and target sheets).
in addition, the old link icon (exportSheet) was replaced by a real link icon
minor renaming and cleaning up of worksheet methods