adding linked sheet highlighting and refactoring icons
when you highlight over a linked sheet icon the corresponding sheet gets an
outline
in order to deal better with css, as well as event handling for span.svg
elements
I made sure that all event were on the parent span, and that
clicking (pointer-events: none) was set on the svg children. Otherwise, there
were touchy issues dealing with propagation which are know for svg's.
Spans that are clickable now have a [data-clickable='true'] attribute and, in
addition, are refernces by ID (ID's have been removed from the SVG icons themselves)