adding file drag and drop
you can now drop a (csv) file onto the Worksheet. It recongnizes
whethere the drop event, is a 'file drop' event, and calls
worksheet.fileUpload() accordingly.
In addition now an overlay appears when you drag over a worksheet. Depending on
what the drag event is a link or a file add icon will appear. For this dragOver
and dragLeave had to be updated and added, respectively.
Minor: the worksheet inherits the name of the original file