Ensuring that the sheet container is properly flexed
-- What
Prior to this, the #sheet-containter shadow dom element had width and
height set to 100%, which means match the parent's height. This is not
correct; we want it to "flex" the dimension, ie calculate out the
header and footer heights.

This is now what happens.

It is a slight (for now) improvement to resizing. The rest will have
to be handled again in apsheet or some combination of both.