adding callstack.onStep() which can be set by a worksheet
Worksheet.callStack.onStep is now set to Worksheet.onCallStackStep
which does some basic highligting of the tab rows. This is just an example of
what we can do and should be worked on further.

adding tests for stepping through the instructions
forcing the worksheet to have source and target set before running through instructions