Adding ContextMenuHandler and initial integration
-- What
This provides an initial integration of the ContextMenuHandler into
the Worksheet. The handler binds/unbinds and handles all specific
events from the ap-sheet or worksheet (or any elements therein). For
example -- and for this initial commit -- we add the ability to have a
context menu on ap-sheet tab clicks, when those clicks are captured at
the ap-sheet level.