moving all connection related logic and control to the ws-connection
all information and logic related to creating, adding, and removing connections
is now handled by ws-connection elements. A given worksheet no longer "knows"
about its sources and targets, ie these attributes are gone, and only updates
corresponding connection elements as needed.

ws-connection elements add and remove leaderlines, and tell the corresponding
worksheet that a target or sources has been added or removed. It is up to the
worksheet what to do with this information (currently it adds and remove
corresponding icons as before).

the current _getInstructions() is a bit hacky since it still assumes a mapper
worksheet as before, which will change.

a bit of cleanup and minor fixes